Back to working state, with changes to allow particles

+++ Vulkan/Game/Camera.cpp
@@ -37,8 +37,8 @@ void Camera::start()
 void Camera::update(float dt)
 {
 	dt = glm::min(dt, MAX_ROTATION_DT);
-	yaw_ += inputManager_->getRelativeMousePos().x * MOUSE_SENSITIVITY * dt; // movement around y axis
-	pitch_ += inputManager_->getRelativeMousePos().y * MOUSE_SENSITIVITY * dt * -1.0f; // movement around x axis
+	yaw_ += static_cast<float>(inputManager_->getRelativeMousePos().x) * MOUSE_SENSITIVITY * dt; // movement around y axis
+	pitch_ += static_cast<float>(inputManager_->getRelativeMousePos().y) * MOUSE_SENSITIVITY * dt * -1.0f; // movement around x axis
 	// clamp pitch to avoid inversion
 	pitch_ = glm::clamp(pitch_, 181.0f, 359.0f);
 	// Calc point on sphere with these (i.e. the two circle's intersection)
