+++ Vulkan/Game/SunScript.h
@@ -23,7 +23,7 @@ namespace QZL {
 		private:
 			float angle_;
 
-			static constexpr float TWO_PI = std::_Pi * 2.0f;
+			static constexpr float TWO_PI = static_cast<float>(std::_Pi) * 2.0f;
 			static constexpr float SPEED = 0.00001f;
 			static constexpr float DISTANCE_PER_SECOND = SPEED * TWO_PI;
 			static constexpr float RADIUS = 81000.0f;
